Resolve conflicts in decode.py (_check_if_req_exceed_kv_capacity) and
model_runner.py (init_routed_experts_capturer / init_indexer_capturer +
max_token_pool_size property) — keep HiSparse host-backed capacity path
on top of main's refactors.

Add unit tests covering:
- ModelRunner.max_token_pool_size HiSparse / SWA / fallback branches
- DecodePreallocQueue._check_if_req_exceed_kv_capacity admission for
  HiSparse (host-backed capacity), non-HiSparse (device pool), and
  rebootstrap length accounting

+++ python/sglang/multimodal_gen/configs/pipeline_configs/sana.py
@@ -52,7 +52,7 @@ class SanaPipelineConfig(SpatialImagePipelineConfig):
     # DC-AE does not support tiling or SP VAE decode yet.
     vae_tiling: bool = False
     vae_sp: bool = False
-    vae_precision: str = "bf16"
+    vae_precision: str = "fp32"
 
     dit_config: DiTConfig = field(default_factory=SanaConfig)
     vae_config: VAEConfig = field(default_factory=SanaVAEConfig)
